+++ plugins/module_utils/network/ios/rm_templates/bgp_address_family.py
@@ -1048,15 +1048,27 @@ class Bgp_address_familyTemplate(NetworkTemplate):
             "name": "as_override",
             "getval": re.compile(
                 r"""
-                \s\sneighbor\s(?P<neighbor_address>\S+)\sas-override
+                \s\sneighbor\s(?P<neighbor_address>\S+)
+                (\s(?P<set>as-override))
+                (\s(?P<split_horizon>split-horizon))?
                 $""",
                 re.VERBOSE,
             ),
             "setval": "neighbor {{ neighbor_address }}"
-            "{{ (' as-override') if as_override|d(False) else '' }}",
+            "{{ (' as-override') if as_override|d(False) else '' }}"
+            "{{ (' split-horizon') if as_override.split_horizon|d(False) else '' }}",
             "result": {
                 "address_family": {
-                    UNIQUE_AFI: {"neighbors": {UNIQUE_NEIB_ADD: {"as_override": True}}},
+                    UNIQUE_AFI: {
+                        "neighbors": {
+                            UNIQUE_NEIB_ADD: {
+                                "as_override": {
+                                    "set": "{{ not not set }}",
+                                    "split_horizon": "{{ not not split_horizon }}",
+                                },
+                            },
+                        },
+                    },
                 },
             },
         },